ORM: Ensure correct error types are returned
The new orm.ErrorConflict was introduced recently to act as a table
signal to callers when conflicts occur, such as record already existing.
It was implemented only within the orm.CreateBridgeType(..) method.

To make the ORM entity more consistent, ensure support for returning
this error is added to all ORM's public methods.